requirements.txt
Flask-Restful
dbbase
Flask-JWT-extended
